Bump version to 1:2.1.0-1kurento1
This package must always supersede official Ubuntu versions, because in
here
we're including a modified version of the library that is required by
Kurento to work properly. If some future update from Ubuntu was
installed
instead of our package, then Kurento won't be able to work with lost
packets.

At the same time, we cannot use an unique package name because libsrtp
is
used not directly by Kurento, but by other system packages
(gstreamer1.0-plugins-bad), so we must adhere to the package name that
is
specified in those system packages.