debian: Depend on exact version of libsrtp
This makes the installation of kurento package to upgrade libsrtp, if it
happened to be already installed in an earlier (system default) version.

Otherwise, upon installation of libsrtp-kurento, the actual libsrtp2-1
will not be upgraded.

Note that debian/control cannot be deleted: kurento-buildpackage expects
it to exist so build dependencies can be installed before attempting to
build the package itself.